requirements.txt
numpy
tensorflow
matplotlib
pandas
scikit-learn
seaborn
conseal
opencv-python
jpeglib
tensorboard
torchinfo
timm